pyproject.toml
[tool.poetry]
name = "attestation"
version = "0.1.0"
description = ""
authors = ["Pvl1307 <ivanov.pavel.1307@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
